koemyd.fetching:socket:oops:reset() server sock on connect() address change
link.server was not being reset() when a subsequent request on the
persistent link contained a different address. although this works
as a oneline-bugfix, i believe a server socket pool could help re-
duce the number of reconnections, so that the link to 192.0.2.1 in

HEAD http://192.0.2.1/ HTTP/1.1
...
HEAD http://192.0.2.2/ HTTP/1.1
...
HEAD http://192.0.2.3/ HTTP/1.1
...
HEAD http://192.0.2.1/ HTTP/1.1
...

would be the same; of course, this may be limited by the timeout
settings of upstream, in which case a reconnect() procedure may
be able to put a new link in this hypothetical "socket pool".

also, added '*.conf' to .gitignore.